+++ collector/processor/concurrentbatchprocessor/testdata/config.yaml
